Added support for CharacterStatSet object creation
Character Stat Sets can now be created by either the "Assets > Create >
Golego > CharacterStatSet" menu or by right clicking the project window
and selecting "Create > Golego > CharacterStatSet". The sets are not
currently implemented anywhere, and are not currently complete. There
will be more fields added to them when the time comes for them.